testimonials crud|home|listing|show car|count category

+++ database/migrations/2024_02_05_214537_create_cars_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -13,12 +12,12 @@ return new class extends Migration
     {
         Schema::create('cars', function (Blueprint $table) {
             $table->id();
-            $table->string('title',100);
+            $table->string('title', 100);
             $table->longText('content');
             $table->integer('luggage');
             $table->integer('doors');
             $table->integer('passengers');
-            $table->float('price');
+            $table->integer('price');
             $table->boolean('active');
             $table->string('image');
             $table->foreignId('cat_id')->constrained('categories');
